% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_eff.R
\name{e_eff}
\alias{e_eff}
\alias{e_eff.default}
\title{Environment Effects}
\usage{
\method{e_eff}{default}(.data, .y, .gen, .env)
}
\arguments{
\item{.data}{data.frame}

\item{.y}{Response Variable}

\item{.gen}{Genotypes Factor}

\item{.env}{Environment Factor}
}
\value{
Environment Effects
}
\description{
Calcuates Environment Effects
}
\examples{

data(Maiz)
e_eff(
    .data  = Maiz
   , .y    = y
   , .gen  = entry
   , .env  = site
   )


}
\references{
Crossa, J., Perez-Elizalde, S., Jarquin, D., Cotes, J.M., Viele, K., Liu, G., and Cornelius, P.L. (2011)
 Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model
 \emph{Crop Science}, \strong{51}, 1458–1469.
 (\href{https://dl.sciencesocieties.org/publications/cs/abstracts/51/4/1458?access=0&view=pdf}{doi: 10.2135/cropsci2010.06.0343})
}
\author{
\enumerate{
    \item Muhammad Yaseen (\email{myaseen208@gmail.com})
   }
}
