% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_var.R
\name{ge_var}
\alias{ge_var}
\alias{ge_var.default}
\title{Genotype by Environment Interaction Variances}
\usage{
\method{ge_var}{default}(.data, .y, .gen, .env)
}
\arguments{
\item{.data}{data.frame}

\item{.y}{Response Variable}

\item{.gen}{Genotypes Factor}

\item{.env}{Environment Factor}
}
\value{
Genotype by Environment Interaction Variances
}
\description{
Calcuates Genotype by Environment Interaction Variances
}
\examples{

data(Maiz)
ge_var(
    .data  = Maiz
   , .y    = y
   , .gen  = entry
   , .env  = site
   )


}
\references{
Crossa, J., Perez-Elizalde, S., Jarquin, D., Cotes, J.M., Viele, K., Liu, G., and Cornelius, P.L. (2011)
 Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model
 \emph{Crop Science}, \strong{51}, 1458–1469.
 (\href{https://acsess.onlinelibrary.wiley.com/doi/abs/10.2135/cropsci2010.06.0343}{doi: 10.2135/cropsci2010.06.0343})
}
\author{
\enumerate{
    \item Muhammad Yaseen (\email{myaseen208@gmail.com})
   }
}
