% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_k.R
\name{matrix_k}
\alias{matrix_k}
\alias{matrix_k.default}
\title{k Matrix}
\usage{
matrix_k(n)

\method{matrix_k}{default}(n)
}
\arguments{
\item{n}{Number of columns}
}
\value{
Matrix
}
\description{
Gives k matrix
}
\references{
Crossa, J., Perez-Elizalde, S., Jarquin, D., Cotes, J.M., Viele, K., Liu, G., and Cornelius, P.L. (2011)
 Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model
 \emph{Crop Science}, \strong{51}, 1458–1469.
 (\href{https://acsess.onlinelibrary.wiley.com/doi/abs/10.2135/cropsci2010.06.0343}{doi: 10.2135/cropsci2010.06.0343})
}
\author{
\enumerate{
    \item Muhammad Yaseen (\email{myaseen208@gmail.com})
   }
}
