% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthnorm.R
\name{orthnorm}
\alias{orthnorm}
\alias{orthnorm.default}
\title{Orthogonal Normalization}
\usage{
orthnorm(u = NULL, basis = TRUE, norm = TRUE)

\method{orthnorm}{default}(u = NULL, basis = TRUE, norm = TRUE)
}
\arguments{
\item{u}{Matrix}

\item{basis}{Logical argument by default TRUE}

\item{norm}{Logical argument by default TRUE}
}
\value{
Matrix
}
\description{
Perform Orthogonal Normalization of a matrix
}
\references{
Crossa, J., Perez-Elizalde, S., Jarquin, D., Cotes, J.M., Viele, K., Liu, G., and Cornelius, P.L. (2011)
 Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model
 \emph{Crop Science}, \strong{51}, 1458–1469.
 (\href{https://acsess.onlinelibrary.wiley.com/doi/abs/10.2135/cropsci2010.06.0343}{doi: 10.2135/cropsci2010.06.0343})
}
\author{
\enumerate{
    \item Muhammad Yaseen (\email{myaseen208@gmail.com})
   }
}
